examples/worker-client/requirements.txt
pico-celery
pico-boot
redis
